moved byebye commands to byebye folder
